requirements.txt
streamlit==1.34.0
tensorflow-cpu==2.12.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
